【提交摘要】优化部分代码结构
