+++ qf_setting.go
